+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* unit tests for `TextDecorator` struct in `src/run.rs`
+
 ### Changed
 * update crate: flood-tide(0.2.14), flood-tide-gen(0.2.2)
 * patch crate: getrandom(0.3.4)